change the ilo Github link part to just be Github
i apologize for how unreadable that title is. all this pr does is it makes the headnoun on *ilo* Github not a link. if anybody can figure out a way to explain this better, please do tell me. i am doing this just for a little more consistency on whether or not headnouns are clickable in the about page, since it seems slightly random to me.